+++ frontend/portfolio-website/scripts/test-integration.bash
@@ -1,0 +1,8 @@
+#!/bin/bash
+# makes sure the folder containing the script will be the root folder
+cd "$(dirname "$0")" || exit
+
+source ../.env
+
+echo 'Running integration tests...'
+echo 'Error: Integration tests are not implemented yet.'