+++ frontend/portfolio-website/scripts/test-unit.bash
@@ -1,0 +1,8 @@
+#!/bin/bash
+# makes sure the folder containing the script will be the root folder
+cd "$(dirname "$0")" || exit
+
+source ../.env
+
+echo 'Running unit tests...'
+echo 'Error: Unit tests are not implemented yet.'