+++ frontend/portfolio-website/src/core/library/atoms/typography/Paragraph.atom.js
@@ -1,0 +1,10 @@
+import PropTypes from "prop-types";
+
+export default function Paragraph(props) {
+  const { classNames = [], text } = props;
+  return (
+    <p className={[...classNames].join(" ")}>
+      {props.children ?? text}
+    </p>
+  );
+}