+++ frontend/portfolio-website/src/pages/_app.js
@@ -1,3 +1,14 @@
-export default function App({ Component, pageProps }) {
-    return <Component {...pageProps} />;
+import "@/styles/base/index.scss";
+
+import { SitePropsProvider } from "@/context/SitePropsContext";
+
+function App({ Component, pageProps }) {
+    return (
+        <SitePropsProvider>
+            <Component {...pageProps} />
+        </SitePropsProvider>
+    );
 }
+
+export default App;
+